refactor news helpers and review cleanup
