+++ src/commands/noodleSocial.js
@@ -51,6 +51,7 @@ import { grantEventBadgesForKnownRecipes } from "../game/badges.js";
 import { hasNewShopLevelSpecialization } from "../game/specialization.js";
 import { isFishingIngredientLocked } from "../game/fishing.js";
 import { nowTs, dayKeyUTC, parseYYYYMMDD } from "../util/time.js";
+import { hasUnreadNewsUpdate } from "../util/news.js";
 import { hasDailyRewardAvailable } from "../game/daily.js";
 import { getOrdersMeta } from "../game/orders.js";
 import { containsProfanity } from "../util/profanity.js";
@@ -991,59 +992,6 @@ function partyCreationRow(userId) {
 /**
  * Social stats view buttons (two rows)
  */
-function normalizeNewsClassification(value) {
-  const raw = String(value ?? "player_update").trim().toLowerCase();
-  if (raw === "internal_update") return "internal_update";
-  return "player_update";
-}
-
-function parseNewsDateMs(value) {
-  const raw = String(value ?? "").trim();
-  if (!raw) return 0;
-  const parsed = Date.parse(raw);
-  return Number.isFinite(parsed) ? parsed : 0;
-}
-
-function normalizeNewsVersion(value) {
-  const raw = String(value ?? "").trim();
-  if (!raw) return "";
-  return raw.startsWith("v") ? raw.toLowerCase() : `v${raw.toLowerCase()}`;
-}
-
-function getLatestNewsVersionForPlayer() {
-  const sections = Array.isArray(newsContent?.sections) ? newsContent.sections : [];
-  const latestVisibleEntry = sections
-    .flatMap((section, sectionIndex) => {
-      const entries = Array.isArray(section?.entries) ? section.entries : [];
-      return entries.map((entry, entryIndex) => ({ entry, sectionIndex, entryIndex }));
-    })
-    .filter(({ entry }) => normalizeNewsClassification(entry?.classification) !== "internal_update")
-    .sort((a, b) => {
-      const diff = parseNewsDateMs(b.entry?.date) - parseNewsDateMs(a.entry?.date);
-      if (diff !== 0) return diff;
-      const sectionDiff = a.sectionIndex - b.sectionIndex;
-      if (sectionDiff !== 0) return sectionDiff;
-      return a.entryIndex - b.entryIndex;
-    })[0] ?? null;
-
-  if (latestVisibleEntry?.entry) {
-    return normalizeNewsVersion(latestVisibleEntry.entry.version);
-  }
-
-  const pinned = newsContent?.pinned ?? {};
-  if (normalizeNewsClassification(pinned?.classification) === "internal_update") {
-    return "";
-  }
-  return normalizeNewsVersion(pinned?.version);
-}
-
-function hasUnreadNewsUpdate(player) {
-  const latestVersion = getLatestNewsVersionForPlayer();
-  if (!latestVersion) return false;
-  const seenVersion = normalizeNewsVersion(player?.notifications?.news_last_seen_version);
-  return seenVersion !== latestVersion;
-}
-
 function statsViewButtons(userId, { newsAvailable = false } = {}) {
   const row1 = new ActionRowBuilder().addComponents(
     new ButtonBuilder()
@@ -1076,7 +1024,7 @@ function statsViewButtons(userId, { newsAvailable = false } = {}) {
     new ButtonBuilder()
       .setCustomId(`noodle:nav:news:${userId}`)
       .setLabel("News")
-      .setEmoji(getButtonEmoji("note"))
+      .setEmoji(getButtonEmoji("new"))
       .setStyle(newsAvailable ? ButtonStyle.Success : ButtonStyle.Secondary),
     new ButtonBuilder()
       .setCustomId(`noodle-social:nav:profile:${userId}`)
@@ -3082,7 +3030,7 @@ async function handleComponent(interaction) {
 
     if (action === "stats") {
       const player = ensurePlayer(serverId, userId);
-      const newsAvailable = hasUnreadNewsUpdate(player);
+      const newsAvailable = hasUnreadNewsUpdate(player, newsContent);
       const tipStats = getUserTipStats(db, serverId, userId);
       const party = getUserActiveParty(db, userId);
       const blessing = getActiveBlessing(player);
